# don't copy byte-array values