This PR is a follow up to https://github.com/segmentio/parquet-go/pull/265 which addresses the performance regression that I had introduced to address the original bug.

I left documentation in the code, at a high level the fix is to never read more than one page per column in each call to `ReadRows`, so we don't invalidate pointers written to the output `[]Row` values.

Contributes to https://github.com/segmentio/parquet-go/issues/226
```
name                                                 old time/op  new time/op  delta
MergeFiles/BYTE_ARRAY/groups=2,rows=20000            51.4µs ± 4%  11.8µs ± 2%   -77.04%  (p=0.000 n=10+10)
MergeFiles/FIXED_LEN_BYTE_ARRAY/groups=2,rows=20000  32.9µs ± 2%   7.6µs ± 1%   -76.92%  (p=0.000 n=10+10)

name                                                 old row/s    new row/s    delta
MergeFiles/BYTE_ARRAY/groups=2,rows=20000             18.8M ± 4%   82.0M ± 2%  +335.29%  (p=0.000 n=10+10)
MergeFiles/FIXED_LEN_BYTE_ARRAY/groups=2,rows=20000   29.4M ± 2%  127.6M ± 1%  +333.30%  (p=0.000 n=10+10)
```